Refactored code (part b).

+++ src/Castle.Facilities.NHibernateIntegration.Tests/Common/Order.cs
@@ -24,6 +24,6 @@ namespace Castle.Facilities.NHibernateIntegration.Tests
 
         public virtual int Id { get; set; }
 
-        public virtual float Value { get; set; }
+        public virtual double Value { get; set; }
     }
 }
